# docs/swagger.yaml
basePath: /
definitions:
  auth.RegisterRequest:
    properties:
      display_name:
        type: string
      email:
        type: string
      password:
        type: string
    type: object
  auth.RegisterResponse:
    properties:
      display_name:
        type: string
      email:
        type: string
      id:
        type: string
    type: object
  auth.RevokeRequest:
    properties:
      token:
        type: string
    type: object
  auth.TokenRequest:
    properties:
      grant_type:
        type: string
      password:
        type: string
      refresh_token:
        type: string
      scope:
        type: string
      username:
        type: string
    type: object
  auth.TokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      refresh_token_expires_in:
        type: integer
      scope:
        type: string
      token_type:
        type: string
    type: object
  posts.CreatePostRequest:
    properties:
      content:
        type: string
      title:
        type: string
    type: object
  posts.PostResponse:
    properties:
      author_id:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      published:
        type: boolean
      title:
        type: string
      updated_at:
        type: string
    type: object
  posts.UpdatePostRequest:
    properties:
      content:
        type: string
      title:
        type: string
    type: object
  response.ErrorPayload:
    properties:
      error:
        type: string
      error_description:
        type: string
    type: object
  users.UpdateProfileRequest:
    properties:
      display_name:
        type: string
    type: object
  users.UserResponse:
    properties:
      created_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      id:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Core reference microservice with full users and posts CRUD.
  title: Reference API
  version: 0.1.0
paths:
  /auth/oauth/revoke:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RevokeRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorPayload'
      summary: Revoke a refresh token
      tags:
      - auth
  /auth/oauth/token:
    post:
      consumes:
      - application/json
      description: Supports password (username, password, optional scope) and refresh_token
        (refresh_token) grants.
      parameters:
      - description: OAuth2 grant payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.TokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorPayload'
      summary: Exchange an OAuth2 token grant
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorPayload'
      summary: Register a user
      tags:
      - auth
  /posts/:
    get:
      parameters:
      - description: Maximum number of posts
        in: query
        name: limit
        type: integer
      - description: Pagination offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/posts.PostResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorPayload'
      summary: List published posts
      tags:
      - posts
    post:
      consumes:
      - application/json
      parameters:
      - description: Post payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/posts.CreatePostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/posts.PostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorPayload'
      security:
      - OAuth2Auth:
        - user
      summary: Create a post
      tags:
      - posts
  /posts/{postID}:
    delete:
      parameters:
      - description: Post UUID
        in: path
        name: postID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorPayload'
      security:
      - OAuth2Auth:
        - user
      summary: Delete a post
      tags:
      - posts
    get:
      parameters:
      - description: Post UUID
        in: path
        name: postID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/posts.PostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorPayload'
      summary: Get a post
      tags:
      - posts
    put:
      consumes:
      - application/json
      parameters:
      - description: Post UUID
        in: path
        name: postID
        required: true
        type: string
      - description: Post payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/posts.UpdatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/posts.PostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorPayload'
      security:
      - OAuth2Auth:
        - user
      summary: Update a post
      tags:
      - posts
  /posts/{postID}/publish:
    post:
      parameters:
      - description: Post UUID
        in: path
        name: postID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/posts.PostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorPayload'
      security:
      - OAuth2Auth:
        - user
      summary: Publish a post
      tags:
      - posts
  /users/me:
    delete:
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorPayload'
      security:
      - OAuth2Auth:
        - user
      summary: Delete the current user
      tags:
      - users
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorPayload'
      security:
      - OAuth2Auth:
        - user
      summary: Get the current user
      tags:
      - users
    put:
      consumes:
      - application/json
      parameters:
      - description: Profile update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/users.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorPayload'
      security:
      - OAuth2Auth:
        - user
      summary: Update the current user
      tags:
      - users
schemes:
- http
- https
securityDefinitions:
  OAuth2Auth:
    flow: password
    scopes:
      user: Access the authenticated user's profile and posts.
    tokenUrl: /auth/oauth/token
    type: oauth2
swagger: "2.0"